classize(), elementize() moved to Config.
This changes the nature of class elements.  While
element IDs are still calculated based on a path
a new map() method allows the app to map a class
to a given configuration file when it's loaded.

Additionally, Config now looks for the class
configuration key and establishes that mapping
on it's own if need be.

You now get configs based on class name.  Core
configs are mapped by inkwell's init method in
the fasion of @config, example:

$config->get('array', '@routes')

Would get the routes configuration.