Modular configs now use @ prefix. IoC added.
When you define the types on a config you can
define it as a core type which is automatically
registered with @ prefix relative to the config
basename.  Example @autoloaders for the core
config /autoloaders.php or @whatever/corecomp for
/whatever/corecomp.php.

Using getAllByType on the config will allow you
to indiciate these types and aggregate all
configs using them.

Controller would not use this as it does not
impart actual core functionality and it is
up to controller as a class to get all
controller configs.  However, controller would
now have a type of @routes and a key of @routes
to signify routes.

The "IoC" container... functionality added.

->register($alias, $class, $factory)
->create($class, $interfaces, $arg1, $arg2)