More route/response/controller work.
The 'response.php' config file was moved down to the root directory and
re-typed as core.  Response class will have API added to register responses
and it will be configured during IW->run() similar to routes.  Rightfully it is
not the same as the rest of the library.

This means inKWell at it's base will simply be IW, Config, Routes, Request, and
Response.  EVERYTHING ELSE IS AN EXTENSION.

This is reflected in other library configs which have been re-typed as
"Extension."

It is decided that configs will *have* to specify a class if they are not core or
Module configs.  This is the only consistent way to ensure proper mapping with
namespaces.  It shouldn't be a pain since scaffolding should make it pretty
straightforward.  The old method for mapping a class at load time still
exists, but will likely be removed and an exception will be thrown if a
configuration exists without a class.

The routing logic was moved to includes/run and Response was given the
ability to only output headers.  Additionally send() now looks at the SAPI
and if it's the console and headers only are being sent, it prints them.

Console (at least for *nix) also now returns the response code.  However,
since exit status codes can only got up to 256, pretty much every response
code maps to (code - 256) -- some alternative mapping may be figured out in
the future.

$_GET is no longer effed with.... period.  Parameters coming from the routes
are added to the request via the ->set() method.

Other minor cleanups.