IW::config() now used Config::load() + added support for 'Library' specific configs.
This is based on the 'Library' config type.  Keys can be overloaded or generated based
on the basename of the class.  But root directories registered are expected to have a
unique reference key.  Additionally a loader instance for the class's __match()
method is added.

Inkwell interface will likely be separated into Init, Matchable, and Buildable
interfaces.