Configs updates for modular aliases and other light cleanup
Aliases were moved to the inkwell config and typecasting argument was made
optional so that config overloading did not have to repeat types.  

The IW loading standard was made to remove underscorizing and focus more on
trait/exception/interface organizations, it is now considered stable and any major
shifts will be for IW2 standard.

Removed IW_CONFIG from .htaccess (this would overload server in a poor way)