Model::search() no longer resolves to default
This fixes issues where the same class can be loaded under almost any namespace
because it will __match().  resolveDatabase() still returns default as it is used only
to resolve for methods like store() and remove().
    
In this sense, the namespace on the database is not strict except for matching which
will prevent models from being scaffolded under various namespaces.